Tweaked toggling scenes to avoid most of the unnecessary flashing
When toggling scenes, the power on and color change commands ran in tandem. This caused a brief flash of color when it was supposed to smoothly toggle the scene.

The recent change lowers the risk of this happening. There is still a bit of a flash, but it is greatly minimized as the power change only occurs after the color settings have been set